Handle sessions with no tracks (e.g. msaview has no view.tracks)
