+++ plugins/menus/src/SessionManager/components/util.ts
@@ -1,0 +1,15 @@
+import { AbstractSessionModel } from '@jbrowse/core/util'
+
+export interface SessionSnap {
+  name: string
+  views?: { tracks?: unknown[] }[]
+  [key: string]: unknown
+}
+
+export interface SessionModel extends AbstractSessionModel {
+  savedSessions: SessionSnap[]
+  removeSavedSession: (arg: SessionSnap) => void
+  activateSession: (arg: string) => void
+  loadAutosaveSession: () => void
+  previousAutosaveId: string
+}
