+++ plugins/variants/src/StructuralVariantChordRenderer/ReactComponent.tsx
@@ -32,30 +32,17 @@ const StructuralVariantChordsReactComponent = observer(function ({
   // make a map of refName -> blockDefinition
   const blocksForRefsMemo = useMemo(() => {
     const blocksForRefs = {} as Record<string, Block>
-    blockDefinitions.forEach(block => {
-      ;(block.region.elided ? block.region.regions : [block.region]).forEach(
-        r => (blocksForRefs[r.refName] = block),
-      )
-    })
+    for (const block of blockDefinitions) {
+      const regions = block.region.elided
+        ? block.region.regions
+        : [block.region]
+      for (const region of regions) {
+        blocksForRefs[region.refName] = block
+      }
+    }
     return blocksForRefs
   }, [blockDefinitions])
-  const chords = []
-  for (const feature of features.values()) {
-    const id = feature.id()
-    const selected = String(selectedFeatureId) === String(id)
-    chords.push(
-      <Chord
-        key={id}
-        feature={feature}
-        config={config}
-        radius={radius}
-        bezierRadius={bezierRadius}
-        blocksForRefs={blocksForRefsMemo}
-        selected={selected}
-        onClick={onChordClick}
-      />,
-    )
-  }
+
   const trackStyleId = `chords-${
     typeof jest !== 'undefined' ? 'test' : displayModel.id
   }`
@@ -72,7 +59,23 @@ const StructuralVariantChordsReactComponent = observer(function ({
 `,
         }}
       />
-      {chords}
+      {[...features.values()].map(feature => {
+        const id = feature.id()
+        const selected = String(selectedFeatureId) === String(id)
+
+        return (
+          <Chord
+            key={id}
+            feature={feature}
+            config={config}
+            radius={radius}
+            bezierRadius={bezierRadius}
+            blocksForRefs={blocksForRefsMemo}
+            selected={selected}
+            onClick={onChordClick}
+          />
+        )
+      })}
     </g>
   )
 })
