+++ website/src/pages/contact.mdx
@@ -2,7 +2,6 @@
 id: contact
 ---
 
-
 import GoogleCalendarScheduleFunction from './util'
 
 # Contact us
@@ -10,7 +9,7 @@ import GoogleCalendarScheduleFunction from './util'
 We **really** love talking to our users. Please reach out with any thoughts you
 have on what we are building!
 
-<GoogleCalendarScheduleFunction/>
+<GoogleCalendarScheduleFunction />
 
 - Ask a question on our discussion board at
   https://github.com/GMOD/jbrowse-components/discussions
